requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
oauthlib==3.1.0
phue==1.1
pyally==1.1.2
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
urllib3==1.26.4
